tidy, touch ups, flatten return statements when possible, move utils to own file
