+++ src/index.ts
@@ -1,6 +1,15 @@
+import {
+  concatBytes,
+  toBytes,
+  parseHexByte,
+  numberToHex,
+  hexToBytes,
+  bytesToHex,
+  utf8ToBytes,
+} from './utils'
 import { Decoded, Input, List } from './types'
 
-// Types exported outside of this package
+// Types exported for easy use
 export { Decoded, Input, List }
 
 /**
@@ -26,8 +35,8 @@ export function encode(input: Input): Uint8Array {
 }
 
 /**
- * Slices a Uint8Array, throws if the slice goes out-of-bounds of the Uint8Array
- * E.g. safeSlice(RLP.utils.hexToBytes('aa'), 1, 2) will throw.
+ * Slices a Uint8Array, throws if the slice goes out-of-bounds of the Uint8Array.
+ * E.g. `safeSlice(RLP.utils.hexToBytes('aa'), 1, 2)` will throw.
  * @param input
  * @param start
  * @param end
@@ -55,12 +64,11 @@ function decodeLength(v: Uint8Array): number {
 function encodeLength(len: number, offset: number): Uint8Array {
   if (len < 56) {
     return Uint8Array.from([len + offset])
-  } else {
-    const hexLength = numberToHex(len)
-    const lLength = hexLength.length / 2
-    const firstByte = numberToHex(offset + 55 + lLength)
-    return Uint8Array.from(hexToBytes(firstByte + hexLength))
   }
+  const hexLength = numberToHex(len)
+  const lLength = hexLength.length / 2
+  const firstByte = numberToHex(offset + 55 + lLength)
+  return Uint8Array.from(hexToBytes(firstByte + hexLength))
 }
 
 /**
@@ -180,113 +188,10 @@ function _decode(input: Uint8Array): Decoded {
   }
 }
 
-// Check if a string is prefixed by 0x
-function isHexPrefixed(str: string): boolean {
-  return str.length >= 2 && str[0] === '0' && str[1] === 'x'
-}
-
-// Removes 0x from a given String
-function stripHexPrefix(str: string): string {
-  if (typeof str !== 'string') {
-    return str
-  }
-  return isHexPrefixed(str) ? str.slice(2) : str
-}
-
-const hexes = Array.from({ length: 256 }, (v, i) => i.toString(16).padStart(2, '0'))
-function bytesToHex(uint8a: Uint8Array): string {
-  // pre-caching chars could speed this up 6x.
-  let hex = ''
-  for (let i = 0; i < uint8a.length; i++) {
-    hex += hexes[uint8a[i]]
-  }
-  return hex
-}
-
-function parseHexByte(hexByte: string): number {
-  if (hexByte.length !== 2) throw new Error('Invalid byte sequence')
-  const byte = Number.parseInt(hexByte, 16)
-  if (Number.isNaN(byte)) throw new Error('Invalid byte sequence')
-  return byte
-}
-
-// Caching slows it down 2-3x
-function hexToBytes(hex: string): Uint8Array {
-  if (typeof hex !== 'string') {
-    throw new TypeError('hexToBytes: expected string, got ' + typeof hex)
-  }
-  if (hex.length % 2) throw new Error('hexToBytes: received invalid unpadded hex')
-  const array = new Uint8Array(hex.length / 2)
-  for (let i = 0; i < array.length; i++) {
-    const j = i * 2
-    array[i] = parseHexByte(hex.slice(j, j + 2))
-  }
-  return array
-}
-
-// Concatenates two Uint8Arrays into one.
-function concatBytes(...arrays: Uint8Array[]): Uint8Array {
-  if (arrays.length === 1) return arrays[0]
-  const length = arrays.reduce((a, arr) => a + arr.length, 0)
-  const result = new Uint8Array(length)
-  for (let i = 0, pad = 0; i < arrays.length; i++) {
-    const arr = arrays[i]
-    result.set(arr, pad)
-    pad += arr.length
-  }
-  return result
-}
-
-// Global symbols in both browsers and Node.js since v11
-// See https://github.com/microsoft/TypeScript/issues/31535
-declare const TextEncoder: any
-declare const TextDecoder: any
-
-function utf8ToBytes(utf: string): Uint8Array {
-  // @ts-ignore
-  return new TextEncoder().encode(utf)
-}
-
-/** Transform an integer into its hexadecimal value */
-function numberToHex(integer: number | bigint): string {
-  if (integer < 0) {
-    throw new Error('Invalid integer as argument, must be unsigned!')
-  }
-  const hex = integer.toString(16)
-  return hex.length % 2 ? `0${hex}` : hex
-}
-
-/** Pad a string to be even */
-function padToEven(a: string): string {
-  return a.length % 2 ? `0${a}` : a
-}
-
 export const utils = {
   bytesToHex,
   hexToBytes,
   utf8ToBytes,
-}
-
-/** Transform anything into a Uint8Array */
-function toBytes(v: Input): Uint8Array {
-  if (v instanceof Uint8Array) return v
-  if (typeof v === 'string') {
-    if (isHexPrefixed(v)) {
-      return hexToBytes(padToEven(stripHexPrefix(v)))
-    } else {
-      return utf8ToBytes(v)
-    }
-  } else if (typeof v === 'number' || typeof v === 'bigint') {
-    if (!v) {
-      return Uint8Array.from([])
-    } else {
-      return hexToBytes(numberToHex(v))
-    }
-  } else if (v === null || v === undefined) {
-    return Uint8Array.from([])
-  } else {
-    throw new Error('invalid type')
-  }
 }
 
 const RLP = { encode, decode }
